refactor: StubOrderService 패키지 이동
- payment.stubrepository -> common.stub
