+++ order/src/test/java/com/food/order/payment/PaymentCancelTest.java
@@ -6,7 +6,7 @@ import com.food.order.payment.mock.MockOrder;
 import com.food.order.payment.mock.MockPayment;
 import com.food.order.payment.mock.MockPaymentLog;
 import com.food.order.payment.mock.MockRequestUser;
-import com.food.order.payment.stubrepository.StubOrderService;
+import com.food.order.common.stub.StubOrderService;
 import com.food.order.payment.stubrepository.StubPaymentLogService;
 import com.food.order.payment.stubrepository.StubPaymentService;
 import com.food.order.payment.stubrepository.StubPointService;
